Decouple Statement and RecordSet from Reader
- The Statement and the RecordSet class are introduced
to improve Reader usage and behaviour

- The Stream filtering is now supported for every object
except when created from SplFileObject.

- Stream Filtering is simplified and improve for Writer class
