+++ tests/StreamFilterTest.php
@@ -13,36 +13,13 @@ use SplTempFileObject;
  */
 class StreamFilterTest extends PHPUnit_Framework_TestCase
 {
-    public function testInitStreamFilterWithWriterStream()
-    {
-        $filter = 'php://filter/write=string.rot13/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        $this->assertTrue($csv->hasStreamFilter('string.rot13'));
-        $this->assertSame(STREAM_FILTER_WRITE, $csv->getStreamFilterMode());
-    }
-
-    public function testInitStreamFilterWithReaderStream()
-    {
-        $filter = 'php://filter/read=string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        $this->assertTrue($csv->hasStreamFilter('string.toupper'));
-        $this->assertSame(STREAM_FILTER_READ, $csv->getStreamFilterMode());
-    }
-
-    public function testInitStreamFilterWithBothStream()
-    {
-        $filter = 'php://filter/string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        $this->assertTrue($csv->hasStreamFilter('string.toupper'));
-        $this->assertSame(STREAM_FILTER_ALL, $csv->getStreamFilterMode());
-    }
-
-    /**
-     * @expectedException LogicException
-     */
     public function testInitStreamFilterWithSplFileObject()
     {
-        Reader::createFromFileObject(new SplFileObject(__DIR__.'/data/foo.csv'))->getStreamFilterMode();
+        $this->assertSame(
+            STREAM_FILTER_READ,
+            Reader::createFromFileObject(new SplFileObject(__DIR__.'/data/foo.csv'))
+                ->getStreamFilterMode()
+        );
     }
 
     public function testappendStreamFilter()
@@ -74,15 +51,6 @@ class StreamFilterTest extends PHPUnit_Framework_TestCase
         $csv->appendStreamFilter('string.toupper');
     }
 
-    /**
-     * @expectedException OutOfBoundsException
-     */
-    public function testSetInvalidStreamFilterMode()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->setStreamFilterMode(34);
-    }
-
     public function testClearAttachedStreamFilters()
     {
         $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
@@ -90,7 +58,7 @@ class StreamFilterTest extends PHPUnit_Framework_TestCase
         $csv->appendStreamFilter('string.rot13');
         $csv->appendStreamFilter('string.toupper');
         $csv->clearStreamFilter();
-        $this->assertFalse($csv->hasStreamFilter('string.rot13'));
+        $this->assertNotContains('JOHN', (string) $csv);
     }
 
     public function testAddMultipleStreamFilter()
@@ -99,39 +67,9 @@ class StreamFilterTest extends PHPUnit_Framework_TestCase
         $csv->appendStreamFilter('string.tolower');
         $csv->prependStreamFilter('string.rot13');
         $csv->appendStreamFilter('string.toupper');
-        $this->assertTrue($csv->hasStreamFilter('string.tolower'));
-        $csv->removeStreamFilter('string.tolower');
-        $this->assertFalse($csv->hasStreamFilter('string.tolower'));
         foreach ($csv as $row) {
             $this->assertSame($row, ['WBUA', 'QBR', 'WBUA.QBR@RKNZCYR.PBZ']);
         }
-    }
-
-    public function testSwithingStreamFilterMode()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('string.toupper');
-        $this->assertSame(STREAM_FILTER_READ, $csv->getStreamFilterMode());
-        $csv->setStreamFilterMode(STREAM_FILTER_WRITE);
-        $this->assertSame(STREAM_FILTER_WRITE, $csv->getStreamFilterMode());
-        foreach ($csv as $row) {
-            $this->assertSame($row, ['john', 'doe', 'john.doe@example.com']);
-        }
-    }
-
-    public function testGetFilterPath()
-    {
-        $csv = Reader::createFromPath(__DIR__.'/data/foo.csv');
-        $csv->appendStreamFilter('string.rot13');
-        $csv->prependStreamFilter('string.toupper');
-        $this->assertFalse($csv->getIterator()->getRealPath());
-    }
-
-    public function testGetFilterPathWithAllStream()
-    {
-        $filter = 'php://filter/string.toupper/resource='.__DIR__.'/data/foo.csv';
-        $csv = Reader::createFromPath($filter);
-        $this->assertFalse($csv->getIterator()->getRealPath());
     }
 
     public function testSetStreamFilterWriterNewLine()
@@ -139,7 +77,6 @@ class StreamFilterTest extends PHPUnit_Framework_TestCase
         stream_filter_register(FilterReplace::FILTER_NAME.'*', FilterReplace::class);
         $csv = Writer::createFromPath(__DIR__.'/data/newline.csv');
         $csv->appendStreamFilter(FilterReplace::FILTER_NAME."\r\n:\n");
-        $this->assertTrue($csv->hasStreamFilter(FilterReplace::FILTER_NAME."\r\n:\n"));
         $csv->insertOne([1, 'two', 3, "new\r\nline"]);
     }
